requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
requests==2.32.5
redis==5.0.7
python-dotenv==1.0.1
pika==1.3.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
pytest==8.3.3
httpx==0.27.2


aio_pika==9.5.7 
aiormq==6.9.2 
multidict==6.7.0 
pamqp==3.3.0 
propcache==0.4.1 
yarl==1.22.0


numpy==2.3.4
pandas==2.3.3 
python-dateutil==2.9.0.post0 
pytz==2025.2 
six==1.17.0 
tzdata==2025.2

duckdb==1.4.1